feat/CIV-284 removed duplications from AssignCaseToUserHandler and Spec
