+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/caseassignment/AbstractAssignCaseToUserHandler.java
@@ -1,0 +1,52 @@
+package uk.gov.hmcts.reform.civil.handler.callback.camunda.caseassignment;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.RequiredArgsConstructor;
+import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
+import uk.gov.hmcts.reform.ccd.client.model.CallbackResponse;
+import uk.gov.hmcts.reform.civil.callback.Callback;
+import uk.gov.hmcts.reform.civil.callback.CallbackHandler;
+import uk.gov.hmcts.reform.civil.callback.CallbackParams;
+import uk.gov.hmcts.reform.civil.enums.CaseRole;
+import uk.gov.hmcts.reform.civil.helpers.CaseDetailsConverter;
+import uk.gov.hmcts.reform.civil.model.CaseData;
+import uk.gov.hmcts.reform.civil.model.IdamUserDetails;
+import uk.gov.hmcts.reform.civil.service.CoreCaseUserService;
+
+import java.util.Map;
+
+import static uk.gov.hmcts.reform.civil.callback.CallbackType.ABOUT_TO_SUBMIT;
+
+@RequiredArgsConstructor
+public abstract class AbstractAssignCaseToUserHandler extends CallbackHandler {
+
+    private final CoreCaseUserService coreCaseUserService;
+    private final CaseDetailsConverter caseDetailsConverter;
+    private final ObjectMapper objectMapper;
+
+    @Override
+    protected Map<String, Callback> callbacks() {
+        return Map.of(
+            callbackKey(ABOUT_TO_SUBMIT), this::assignSolicitorCaseRole
+        );
+    }
+
+    private CallbackResponse assignSolicitorCaseRole(CallbackParams callbackParams) {
+        CaseData caseData = caseDetailsConverter.toCaseData(callbackParams.getRequest().getCaseDetails());
+        String caseId = caseData.getCcdCaseReference().toString();
+        IdamUserDetails userDetails = caseData.getApplicantSolicitor1UserDetails();
+        String submitterId = userDetails.getId();
+        String organisationId = caseData.getApplicant1OrganisationPolicy().getOrganisation().getOrganisationID();
+
+        coreCaseUserService.assignCase(caseId, submitterId, organisationId, CaseRole.APPLICANTSOLICITORONE);
+        coreCaseUserService.removeCreatorRoleCaseAssignment(caseId, submitterId, organisationId);
+
+        CaseData updated = caseData.toBuilder()
+            .applicantSolicitor1UserDetails(IdamUserDetails.builder().email(userDetails.getEmail()).build())
+            .build();
+
+        return AboutToStartOrSubmitCallbackResponse.builder()
+            .data(updated.toMap(objectMapper))
+            .build();
+    }
+}
